feat: Complete platform split with proper authentication
- Split accounts views into consumer and owner specific modules
- Create separate URL configurations for each platform
- Fix consumer API endpoints to require proper authentication
- Add AllowUserTypes permission to enforce platform separation
- Update consumer views to use IsAuthenticated + AllowUserTypes instead of AllowAny
- Wire up proper URL routing for both platforms
- Test and verify 403 responses for cross-platform access

Platform separation now working correctly:
- Owner tokens → Consumer API: 403 Forbidden ✅
- Consumer tokens → Owner API: 403 Forbidden ✅
- Owner tokens → Owner API: 200 OK ✅
- Consumer tokens → Consumer API: 200 OK ✅
