+++ stores/consumer_views.py
@@ -1,27 +1,28 @@
 # stores/consumer_views.py
 
 from rest_framework import generics
-from rest_framework.permissions import AllowAny
+from rest_framework.permissions import AllowAny, IsAuthenticated
 
+from accounts.permissions import AllowUserTypes
 from .models import Menu, Store
 from .serializers import MenuSerializer, StoreListSerializer
 
 
-# API view to get a list of all stores (public)
+# API view to get a list of all stores (consumer only)
 class StoreListView(generics.ListAPIView):
     """
     Retrieves a list of all stores.
     """
-    permission_classes = [AllowAny]
+    permission_classes = [IsAuthenticated, AllowUserTypes]
     queryset = Store.objects.all()
     serializer_class = StoreListSerializer
 
-# API view to get the menu for a specific store (public)
+# API view to get the menu for a specific store (consumer only)
 class MenuListView(generics.ListAPIView):
     """
     Retrieves the menu list for a specific store.
     """
-    permission_classes = [AllowAny]
+    permission_classes = [IsAuthenticated, AllowUserTypes]
     serializer_class = MenuSerializer
 
     def get_queryset(self):
